Add wrapper for scheduleEntries logic
