+++ frontend/src/views/camp/CampProgram.vue
@@ -19,46 +19,29 @@ Show all activity schedule entries of a single period.
         <v-icon v-if="listFormat">mdi-calendar-month</v-icon>
         <v-icon v-else>mdi-menu</v-icon>
       </v-btn>
-      <template v-if="!(period() && !period()._meta.loading)">
-        <v-skeleton-loader v-if="listFormat" type="list-item-avatar-two-line@2" class="py-2" />
-        <v-skeleton-loader v-else type="table" />
-      </template>
-      <template v-if="firstPeriodLoaded">
-        <picasso
-          v-show="!listFormat"
-          class="mx-2 ma-sm-0 pa-sm-2"
-          :schedule-entries="scheduleEntries"
-          :period="period"
-          :start="Date.parse(period().start)"
-          :end="Date.parse(period().end)"
-          :dialog-activity-create="showActivityCreateDialog"
-          :dialog-activity-edit="showActivityEditDialog" />
-        <activity-list
-          v-show="listFormat"
-          :schedule-entries="scheduleEntries"
-          :period="period" />
-      </template>
-      <dialog-activity-create
-        ref="dialogActivityCreate"
-        :schedule-entry="popupEntry"
-        @activityCreated="afterCreateActivity($event)"
-        @creationCanceled="cancelNewActivity" />
-      <dialog-activity-edit
-        ref="dialogActivityEdit"
-        :schedule-entry="popupEntry" />
-      <v-btn
-        :fixed="$vuetify.breakpoint.xs"
-        :absolute="!$vuetify.breakpoint.xs"
-        dark
-        fab
-        style="z-index: 3"
-        bottom
-        right
-        class="fab--bottom_nav float-right"
-        color="red"
-        @click.stop="createNewActivity()">
-        <v-icon>mdi-plus</v-icon>
-      </v-btn>
+      <schedule-entries :period="period" :showButton="true">
+        <template v-slot:default="slotProps">
+          <template v-if="slotProps.loading">
+            <v-skeleton-loader v-if="listFormat" type="list-item-avatar-two-line@2" class="py-2" />
+            <v-skeleton-loader v-else type="table" />
+          </template>
+          <template v-else>
+            <picasso
+              v-show="!listFormat"
+              class="mx-2 ma-sm-0 pa-sm-2"
+              :schedule-entries="slotProps.scheduleEntries"
+              :period="period"
+              :start="Date.parse(period().start)"
+              :end="Date.parse(period().end)"
+              :dialog-activity-create="slotProps.showActivityCreateDialog"
+              :dialog-activity-edit="slotProps.showActivityEditDialog" />
+            <activity-list
+              v-show="listFormat"
+              :schedule-entries="slotProps.scheduleEntries"
+              :period="period" />
+          </template>
+        </template>
+      </schedule-entries>
     </v-sheet>
   </content-card>
 </template>
@@ -67,9 +50,7 @@ import ContentCard from '@/components/layout/ContentCard'
 import SearchMobile from '@/components/navigation/SearchMobile'
 import Picasso from '@/components/camp/Picasso'
 import ActivityList from '@/components/camp/ActivityList'
-import DialogActivityCreate from '@/components/dialog/DialogActivityCreate'
-import DialogActivityEdit from '@/components/dialog/DialogActivityEdit'
-import { defineHelpers } from '@/components/scheduleEntry/dateHelperLocal'
+import ScheduleEntries from '@/components/scheduleEntry/ScheduleEntries'
 
 export default {
   name: 'CampProgram',
@@ -78,100 +59,18 @@ export default {
     SearchMobile,
     Picasso,
     ActivityList,
-    DialogActivityCreate,
-    DialogActivityEdit
+    ScheduleEntries
   },
   props: {
     period: { type: Function, required: true }
   },
-  data () {
-    return {
-      scheduleEntries: [],
-      deleteTempEntryCallback: () => {},
-      popupEntry: {}
-    }
-  },
   computed: {
     listFormat () {
       return this.$route.query.list
-    },
-    camp () {
-      return this.period().camp()
-    },
-    firstPeriodLoaded () {
-      return this.period() && !this.period()._meta.loading
-    },
-    apiScheduleEntries () {
-      return this.period().scheduleEntries()
     }
-  },
-  watch: {
-    apiScheduleEntries (value) {
-      this.scheduleEntries = value.items.map(entry => defineHelpers(entry, true))
-    }
-  },
-  beforeMount () {
-    this.scheduleEntries = this.apiScheduleEntries.items.map(entry => defineHelpers(entry, true))
-  },
-  methods: {
-    createNewActivity () {
-      const entry = defineHelpers({
-        number: null,
-        period: () => (this.period)(),
-        periodOffset: 420,
-        length: 60,
-        activity: () => ({
-          title: this.$tc('entity.activity.new'),
-          location: '',
-          camp: (this.period)().camp,
-          activityCategory: () => ({
-            id: null,
-            short: null,
-            color: 'grey elevation-4 v-event--temporary'
-          })
-        })
-      }, true)
-      this.showActivityCreateDialog(entry, () => {})
-    },
-    showActivityCreateDialog (entry, deleteTempEntryCallback) {
-      this.popupEntry = entry
-      this.deleteTempEntryCallback = deleteTempEntryCallback
-      this.$refs.dialogActivityCreate.showDialog = true
-    },
-    showActivityEditDialog (entry) {
-      this.popupEntry = entry
-      this.$refs.dialogActivityEdit.showDialog = true
-    },
-    afterCreateActivity (data) {
-      this.api.reload(this.period().scheduleEntries())
-      this.scheduleEntries.push(...data.scheduleEntries().items.map(entry => defineHelpers(entry, true)))
-      this.deleteTempEntryCallback()
-    },
-    cancelNewActivity () {
-      this.deleteTempEntryCallback()
-    },
-    defineHelpers
   }
 }
 </script>
-
-<style lang="scss">
-  .fab--bottom_nav {
-    position: fixed;
-    bottom: 16px + 56px !important;
-    @media #{map-get($display-breakpoints, 'sm-and-up')}{
-      bottom: 16px + 36px !important;
-    }
-  }
-
-  .fab--top_nav {
-    position: fixed;
-    top: 16px + 105px !important;
-    @media #{map-get($display-breakpoints, 'sm-and-up')}{
-      top: 16px + 65px !important;
-    }
-  }
-</style>
 
 <style lang="scss" scoped>
   ::v-deep .v-skeleton-loader__list-item-avatar-two-line {
